refucter:コンポーネントをリファクタリング
